+++ src/components/Form.js
@@ -2,71 +2,83 @@ import React, { Component } from "react";
 
 import classnames from "../classnames";
 
-import Button from "./Button";
+import SubmitButton from "./SubmitButton";
+import CentsField from "./CentsField";
+import { EmailField, NumberField, PasswordField, StringField } from "./FormFields";
+
+const contains = haystack => needle => haystack.indexOf(needle) > -1;
+const isField = contains([CentsField, EmailField, NumberField, PasswordField, StringField]);
 
 class Form extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
+      submitted: false,
       submitting: false,
-      touched: false,
-      values: props.initialValues || {}
+      values: props.initialValues || {},
+      errors: {}
     };
   }
 
-  getIsValidSubmission = () => {
+  getChildren = () => {
     const { children } = this.props;
-    const { values } = this.state;
+    const { submitted, submitting, values } = this.state;
 
-    return children.every(({ props: { required, name } }) => (
-      !required || values[name]
-    ));
+    return React.Children.map(children, child => {
+      const { type, props } = child;
+
+      if (isField(type)) {
+        return React.cloneElement(child, {
+          onError: this.handleError,
+          onFormChange: this.handleFormChange,
+          submitted,
+          value: values[props.name]
+        });
+      }
+
+      if (type === SubmitButton) {
+        return React.cloneElement(child, { submitting });
+      }
+
+      return child;
+    });
   };
 
-  handleValueChange = mutation => {
-    this.setState(({ values }) => ({ values: { ...values, ...mutation } }));
+  handleError = (name, error) => {
+    this.setState(({ errors }) => ({
+      errors: { ...errors, [name]: error }
+    }));
+  };
+
+  handleFormChange = (name, value) => {
+    this.setState(({ values }) => ({
+      values: { ...values, [name]: value }
+    }));
   };
 
   handleSubmit = event => {
     const { onSubmit } = this.props;
-    const { values } = this.state;
+    const { errors, values } = this.state;
 
     event.preventDefault();
 
-    if (!this.getIsValidSubmission()) {
-      this.setState({ touched: true });
-      return false;
+    if (Object.keys(errors).every(name => !errors[name])) {
+      this.setState({ submitting: true });
+
+      const doneSubmitting = () => this.setState({ submitting: false });
+      onSubmit(values).then(doneSubmitting).catch(doneSubmitting);
     }
 
-    this.setState({ submitting: true });
-    const doneSubmitting = () => this.setState({ submitting: false });
-
-    onSubmit(values).then(doneSubmitting).catch(doneSubmitting);
-    return true;
+    this.setState({ submitted: true });
   };
 
   render() {
-    const { children, className, initialValues } = this.props;
-    const { submitting, touched } = this.state;
+    const { className } = this.props;
 
     return (
       <form className={classnames(className)} onSubmit={this.handleSubmit}>
-        {React.Children.map(children, child => (
-          React.cloneElement(child, {
-            onValueChange: this.handleValueChange,
-            initialValues,
-            touched
-          })
-        ))}
-        <Button
-          primary
-          type="submit"
-          disabled={submitting}
-          onClick={this.handleSubmit}
-        >
-          Submit
-        </Button>
+        {this.getChildren()}
       </form>
     );
   }